Moved the lobby parsing over to LobbyHelper
