+++ LobbyCompatibility/Patches/LobbyDataIsJoinablePostfix.cs
@@ -47,7 +47,7 @@ internal static class LobbyDataIsJoinablePostfix
         }
 
         var matchesPluginRequirements =
-            PluginHelper.MatchesTargetRequirements(PluginHelper.ParseLobbyPluginsMetadata(lobbyPluginString));
+            PluginHelper.MatchesTargetRequirements(LobbyHelper.ParseLobbyPluginsMetadata(lobbyPluginString));
 
         if (!matchesPluginRequirements)
         {
